WIP Extract Copyable base class, inheit optGraph from it
